refactor: simplify code
