+++ sgxvm/pkg/sumdb/sum.golang.org/latest
@@ -1,0 +1,5 @@
+go.sum database tree
+23218161
+Q90ELLUdpDqOBNQ8Y7ByMuiknJCeKhrvdETfbaV+wEY=
+
+— sum.golang.org Az3grhvP/nrdYt4pSZWGvB9dS+YpLSkY9NpF5lK2nQFzXxvdC+VJ7B01du7a3LMT8uUmEezO0mFYcmO1A8S7en4uPgA=
